Separate assembly serialisers into components.
